requirements.txt
Flask==2.3.3 # Ou a versão mais recente que preferir
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==2.3.7 # Para hashing de senhas, geralmente vem com Flask
python-dotenv==1.0.0 # Para carregar variáveis de ambiente
# ... outras libs
Flask-Moment==1.2.0 # Ou a versão mais recente